+++ apps/docs/components/ui/pre/Pre.tsx
@@ -1,6 +1,10 @@
-import type { HTMLAttributes } from "react";
+"use client";
 
-import CopyButton from "@/components/copyButton/CopyButton";
+import React from "react";
+import type { HTMLAttributes } from "react";
+import cx from "classnames";
+
+import CodeBlockCopyButton from "@/components/copyButton/codeblockCopyButton/CodeBlockCopyButton";
 import LangIcon from "@/components/ui/pre/langIcon/LangIcon";
 
 import "./pre.css";
@@ -11,8 +15,18 @@ export type PreProps = React.DetailedHTMLProps<HTMLAttributes<HTMLPreElement>, H
 };
 
 const Pre = ({ children, title, "data-language": dataLanguage, raw, ...props }: PreProps) => {
+    let preClasses = "";
+
+    if (!title) {
+        preClasses = "hd-pre--no-title";
+    }
+
+    const classes = cx("hd-pre", preClasses);
+
     return (
-        <pre {...props}>
+        <pre {...props} className={classes}>
+            {raw && <CodeBlockCopyButton className="hd-pre-copy-button hd-copy-button--on-dark" text={raw} />}
+            {title &&
             <div className="hd-pre-header">
                 <div className="hd-pre-header__info">
                     {dataLanguage && (
@@ -22,8 +36,7 @@ const Pre = ({ children, title, "data-language": dataLanguage, raw, ...props }: 
                     }
                     <span className="hd-pre-header__title">{title}</span>
                 </div>
-                {raw && <CopyButton text={raw} className="hd-copy-button--on-dark" />}
-            </div>
+            </div>}
             {children}
         </pre>
     );