+++ packages/components/tsup.build.ts
@@ -1,7 +1,5 @@
 import { defineBuildConfig } from "@workleap/tsup-configs";
 
 export default defineBuildConfig({
-    entry: ["./src/**/src/*.(ts|tsx)"],
-    minify: true,
-    splitting: false
+    entry: ["./src/**/src/*.(ts|tsx)"]
 });