+++ packages/styled-system/tsup.dev.ts
@@ -3,7 +3,7 @@ import { defineDevConfig } from "@workleap/tsup-configs";
 import packageJson from "./package.json";
 
 export default defineDevConfig({
-    entry: ["./src/*.(ts|tsx)"],
+    entry: ["./src/**/*.(ts|tsx)"],
     esbuildPlugins: [
         createCssModuleEsbuildPlugin({
             generateScopedName: "[name]__[local]___[hash:base64:5]",