+++ packages/components/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @hopper-ui/components
 
+## 0.0.4
+
+### Patch Changes
+
+- f40a266: Optimized the build output
+- Updated dependencies [f40a266]
+  - @hopper-ui/styled-system@0.2.2
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [a31b7e2]
+  - @hopper-ui/styled-system@0.2.1
+
 ## 0.0.2
 
 ### Patch Changes