+++ packages/styled-system/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hopper-ui/styled-system
 
+## 0.2.2
+
+### Patch Changes
+
+- f40a266: Optimized the build output
+
+## 0.2.1
+
+### Patch Changes
+
+- a31b7e2: Updated the heading small line height token to the right core token
+
 ## 0.2.0
 
 ### Minor Changes